requirements.txt
Django==3.2
djangorestframework==3.13.1
pylint==2.12.2
psycopg2-binary==2.9.3
python-dotenv==0.19.2
PyJWT==2.3.0
